// frontend/src/api/generated/schemas/shipmentCreateRequestStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SecureShipBackend
 * BFF for SecureShip client application
 * OpenAPI spec version: 1.0.0
 */

export type ShipmentCreateRequestStatus = typeof ShipmentCreateRequestStatus[keyof typeof ShipmentCreateRequestStatus];


export const ShipmentCreateRequestStatus = {
  label_created: 'label_created',
  in_transit: 'in_transit',
  out_for_delivery: 'out_for_delivery',
  delivered: 'delivered',
  exception: 'exception',
} as const;